Move excel CursorWriter to separate module
